+++ examples/plugin/Hello_world/README.md
@@ -12,16 +12,16 @@ It is neccesary to have the repo [ign-plugin](https://github.com/ignitionrobotic
 * CMakeLists.txt
 
 ``` bash 
- cd Hello_world
+ cd hello_world
 mkdir build/
 cd build/
 cmake ..
 make
 ```
 
-It should generate a .so file 
+It should generate a _libHelloWorld.so_ file 
 we can import it to ignition gazebo:
 ``` bash 
-cd Hello_world
+cd hello_world
 export IGN_GAZEBO_SYSTEM_PLUGIN_PATH=`pwd`/build
 ```
